moved SteppedForm.js to common folder removed old form fixed tests
