+++ services/minespace-web/src/components/common/Step.js
